moved blockchain construction to main
